Change more filtering to be done by Prisma
