+++ src/campaigns/services/campaigns.service.ts
@@ -2,7 +2,7 @@ import { BadRequestException, Injectable, Logger } from '@nestjs/common'
 import { PrismaService } from '../../prisma/prisma.service'
 import { UpdateCampaignSchema } from '../schemas/updateCampaign.schema'
 import { CampaignListSchema } from '../schemas/campaignList.schema'
-import { Campaign, Prisma, User } from '@prisma/client'
+import { Campaign, Prisma, Race, User } from '@prisma/client'
 import { deepMerge } from 'src/shared/util/objects.util'
 import { caseInsensitiveCompare } from 'src/prisma/util/json.util'
 import { findSlug } from 'src/shared/util/slug.util'
@@ -17,9 +17,11 @@ import {
   CampaignDetails,
   CampaignLaunchStatus,
   OnboardingStep,
+  CampaignUpdate,
 } from '../campaigns.types'
 import { EmailService } from 'src/email/email.service'
 import { zipToLatLng } from '../util/zipToLatLng'
+import { RaceData } from 'src/races/races.types'
 
 const APP_BASE = process.env.CORS_ORIGIN as string
 
@@ -121,7 +123,7 @@ export class CampaignsService {
         },
         data: {
           slug,
-          currentStep: 'registration',
+          currentStep: OnboardingStep.registration,
         },
       },
     })
@@ -363,428 +365,8 @@ export class CampaignsService {
       this.logger.error('Error sending campaign launch email', e)
     }
   }
-
-  async listMapCount(
-    state?: string,
-    results?: boolean,
-  ): Promise<{ count: number }> {
-    // Config to check APP_BASE ?
-
-    // let whereClauses = `WHERE c."user_id" IS NOT NULL AND c."isDemo" = false AND c."isActive" = true`
-
-    // if (state) {
-    //   whereClauses += ` AND c.details->>'state' = '${state}`
-    // }
-
-    // if (results) {
-    //   whereClauses += ` AND (c."didWin" = true OR c.data->'hubSpotUpdates'->>'election_results' = 'Won General')`
-    // }
-
-    // const rawQuery = `
-    //   SELECT
-    //       c."slug",
-    //       c."details",
-    //       c."didWin",
-    //       c."data",
-    //       u."firstName",
-    //       u."lastName",
-    //       u."avatar"
-    //     FROM "campaign" c
-    //     JOIN "user" u ON c."user" = u.id
-    //     ${whereClauses};
-    //   `
-
-    // const campaigns = await this.prismaService.$queryRawUnsafe<
-    //   {
-    //     slug: string
-    //     details: any
-    //     didWin: boolean | null
-    //     data: any
-    //     firstName: string
-    //     lastName: string
-    //     avatar: string
-    //   }[]
-    // >(rawQuery)
-
-    const whereClause: Prisma.CampaignWhereInput = {
-      userId: { not: null },
-      isDemo: false,
-      isActive: true,
-      ...(state
-        ? {
-            details: {
-              path: ['state'],
-              equals: state,
-            },
-          }
-        : {}),
-      ...(results
-        ? {
-            OR: [
-              { didWin: true },
-              {
-                data: {
-                  path: ['hubSpotUpdates', 'election_results'],
-                  equals: 'Won General',
-                },
-              },
-            ],
-          }
-        : {}),
-    }
-
-    const campaigns = await this.prismaService.campaign.findMany({
-      where: whereClause,
-      select: {
-        slug: true,
-        details: true,
-        didWin: true,
-        data: true,
-        user: {
-          select: {
-            firstName: true,
-            lastName: true,
-            avatar: true,
-          },
-        },
-      },
-    })
-
-    const lastWeek = new Date()
-    lastWeek.setDate(lastWeek.getDate() - 7)
-    let count = 0
-
-    for (const campaign of campaigns) {
-      const { didWin, slug } = campaign
-      const details = campaign.details as CampaignDetailsContent
-      const data = campaign.data as CampaignDataContent
-
-      if (
-        !details?.zip ||
-        didWin === false ||
-        !details?.geoLocation?.lng ||
-        details?.geoLocationFailed
-      ) {
-        if (slug === 'tego sol victoria caelum vestrum') {
-          console.log('Failed first check')
-        }
-        continue
-      }
-      const isProd = false // Change this later
-
-      if (isProd) {
-        if (data?.hubSpotUpdates?.verified_candidates !== 'Yes') {
-          if (slug === 'tego sol victoria caelum vestrum') {
-            console.log('Failed second check')
-          }
-          continue
-        }
-      }
-
-      if (didWin == null) {
-        if (!details.electionDate) {
-          continue
-        }
-        const electionDate = new Date(details.electionDate)
-        if (electionDate < lastWeek) {
-          if (slug === 'tego sol victoria caelum vestrum') {
-            console.log('Failed fourth check')
-          }
-          continue
-        }
-      }
-
-      count++
-    }
-
-    return { count }
-  }
-
-  async listMap(
-    party?: string,
-    state?: string,
-    level?: string,
-    results?: string,
-    office?: string,
-    name?: string,
-    forceReCalc?: boolean,
-  ): Promise<CleanCampaign[]> {
-    // Logic for checking appbase?
-
-    const andConditions: any = []
-
-    if (party) {
-      andConditions.push({
-        details: {
-          path: ['party'],
-          contains: party.toLowerCase(),
-        },
-      })
-    }
-
-    if (state) {
-      andConditions.push({
-        details: {
-          path: ['state'],
-          equals: state,
-        },
-      })
-    }
-
-    if (level) {
-      andConditions.push({
-        details: {
-          path: ['ballotLevel'],
-          equals: level,
-        },
-      })
-    }
-
-    if (results) {
-      andConditions.push({
-        OR: [
-          { didWin: true },
-          {
-            data: {
-              path: ['hubSpotUpdates', 'election_results'],
-              equals: 'Won General',
-            },
-          },
-        ],
-      })
-    }
-
-    if (office) {
-      andConditions.push({
-        OR: [
-          {
-            details: {
-              path: ['normalizedOffice'],
-              equals: office,
-            },
-          },
-          {
-            details: {
-              path: ['office'],
-              equals: office,
-            },
-          },
-          {
-            details: {
-              path: ['otherOffice'],
-              equals: office,
-            },
-          },
-        ],
-      })
-    }
-
-    const isProd = false
-
-    if (isProd) {
-      andConditions.push({
-        data: {
-          path: ['hubSpotUpdates', 'verified_candidates'],
-          equals: 'Yes',
-        },
-      })
-    }
-
-    const where: Prisma.CampaignWhereInput = {
-      userId: { not: null },
-      isDemo: false,
-      isActive: true,
-      AND: andConditions,
-    }
-
-    const campaigns = await this.prismaService.campaign.findMany({
-      where,
-      include: {
-        user: {
-          select: {
-            firstName: true,
-            lastName: true,
-            avatar: true,
-          },
-        },
-      },
-    })
-
-    const cleanCampaigns: CleanCampaign[] = []
-    const lastWeek = new Date()
-    lastWeek.setDate(lastWeek.getDate() - 7)
-
-    for (const campaign of campaigns) {
-      const { didWin, slug, user } = campaign
-      const details = campaign.details as CampaignDetailsContent
-      const data = campaign.data as CampaignDataContent
-
-      if (!details?.zip || didWin === false) {
-        continue
-      }
-
-      const {
-        otherOffice,
-        office,
-        state,
-        ballotLevel,
-        zip,
-        party,
-        raceId,
-        noNormalizedOffice,
-        county,
-        city,
-      } = details || {}
-
-      const electionDate = details.electionDate as string
-
-      const resolvedOffice = (otherOffice as string) || (office as string)
-      if (name) {
-        const fullName = [user?.firstName ?? '', user?.lastName ?? '']
-          .join(' ')
-          .trim()
-          .toLowerCase()
-        if (!fullName.includes(name.toLowerCase())) continue
-      }
-
-      let normalizedOffice =
-        data?.hubSpotUpdates?.office_type || details?.normalizedOffice
-
-      if (!normalizedOffice && raceId && !noNormalizedOffice) {
-        const race = await this.prismaService.race.findFirst({
-          where: { ballotHashId: raceId },
-        })
-        if (
-          typeof race?.data === 'object' &&
-          !Array.isArray(race?.data) &&
-          typeof race?.data?.normalized_position_name === 'string' // Come back to this hot mess
-        ) {
-          normalizedOffice = race?.data?.normalized_position_name
-        }
-        if (normalizedOffice) {
-          await this.prismaService.campaign.update({
-            where: { slug },
-            data: { details: { ...details, normalizedOffice } },
-          })
-        } else {
-          await this.prismaService.campaign.update({
-            where: { slug },
-            data: { details: { ...details, noNormalizedOffice: true } },
-          })
-        }
-      }
-
-      const cleanCampaign: CleanCampaign = {
-        slug,
-        id: slug,
-        didWin,
-        office: resolvedOffice,
-        state: state || null,
-        ballotLevel: ballotLevel || null,
-        zip: zip || null,
-        party: party || null,
-        firstName: campaign.user?.firstName || '',
-        lastName: campaign.user?.lastName || '',
-        avatar: campaign.user?.avatar || false,
-        electionDate,
-        county: county || null,
-        city: city || null,
-        normalizedOffice: normalizedOffice || resolvedOffice,
-      }
-
-      if (didWin === null) {
-        const date = new Date(electionDate)
-        if (date < lastWeek) {
-          continue
-        }
-      }
-
-      const globalPosition = await handleGeoLocation(
-        campaign as Campaign,
-        forceReCalc,
-        this.prismaService,
-      )
-      if (!globalPosition) {
-        continue
-      } else {
-        cleanCampaign.globalPosition = globalPosition
-      }
-
-      cleanCampaigns.push(cleanCampaign)
-    }
-  }
 }
 
-async function handleGeoLocation(
-  campaign: Campaign,
-  forceReCalc: boolean | undefined,
-  prismaService: PrismaService,
-): Promise<{ lat: number; lng: number } | null> {
-  const details = campaign.details as CampaignDetailsContent
-  const { geoLocationFailed, geoLocation } = details || {}
-
-  if (!forceReCalc && geoLocationFailed) {
-    return null
-  }
-
-  if (forceReCalc || !geoLocation?.lng) {
-    const geoLocation = await calculateGeoLocation(campaign, prismaService)
-    if (!geoLocation) {
-      await prismaService.campaign.update({
-        where: {
-          slug: campaign.slug,
-        },
-        data: {
-          details: {
-            ...details,
-            geoLocationFailed: true,
-          },
-        },
-      })
-      return null
-    }
-    return { lng: geoLocation.lng, lat: geoLocation.lat }
-  } else if (geoLocation?.lng && geoLocation?.lat) {
-    return {
-      lng: geoLocation?.lng,
-      lat: geoLocation?.lat,
-    }
-  } else return null
-}
-
-async function calculateGeoLocation(
-  campaign: Campaign,
-  prismaService: PrismaService,
-): Promise<{ lat: number; lng: number; geoHash: string } | null> {
-  if (!campaign.details?.zip || !campaign.details?.state) {
-    return null
-  }
-  const globalCoords = await zipToLatLng(
-    campaign.details?.zip,
-    campaign.details?.state,
-  )
-  if (globalCoords == null) {
-    return null
-  }
-  const { lat, lng, geoHash } = globalCoords
-  await prismaService.campaign.update({
-    where: {
-      slug: campaign.slug,
-    },
-    data: {
-      details: {
-        ...(campaign.details as CampaignDetailsContent),
-        geoLocationFailed: false,
-        geoLocation: {
-          geoHash,
-          lat,
-          lng,
-        },
-      },
-    },
-  })
-  return { lng, lat, geoHash }
-}
 // async function updatePathToVictory(campaign, columnKey, value) {
 //   try {
 //     const p2v = await PathToVictory.findOrCreate(
